feat(cli-commands): complete remaining CLI fix plan phases — security, paths, exit codes
Complete Phase 1 (security), Phase 2 (critical bugs), Phase 3 (performance),
Phase 5 (error handling), and Phase 6 (catalog consolidation) across the
surface layer commands.

Key changes:
- surface_setup_command.rs: compile-time trusted doc bundling via include_str!
  for supply-chain security (P1.3) — SKILL.md, ARCHITECTURE.md, MIGRATION
  guides, RULES_AES bundled at compile time instead of copying from project
- surface_config_command.rs: secret redaction for config show output (P5.2)
  — detects AWS access key IDs (AKIA pattern) and base64-like strings
- surface_maintenance_command.rs: proper exit codes for security tool
  missing (exit code 3 instead of SUCCESS) (P5.1)
- utility_file.rs: canonical path-based file walking, added common ignore
  directories (.git, dist, build, coverage, .venv) (P3.2)
- root_cli_main_entry.rs + taxonomy_catalog_constant.rs: removed Duplicates
  command from CLI and catalog (Phase 6 consolidation)
- surface_git_command.rs: user-provided path + filter support (P2.5)
- surface_check_action.rs: pass user path/filter to git-diff handler (P2.5)
- surface_common_command.rs: float comparison for CI threshold (P2.7)
- surface_plugin_command.rs: dynamic adapter discovery from aggregate

+++ crates/cli-commands/src/surface_common_command.rs
@@ -66,7 +66,8 @@ pub fn run_ci_analysis(
     let results = code_analysis_linter.run_code_analysis_path(&root);
     let score = code_analysis_linter.calc_score(&results);
     let has_crit = code_analysis_linter.check_critical(&results);
-    let below_threshold = (score.value() as u32) < threshold.value();
+    // P2.7: compare as floats, not truncated u32
+    let below_threshold = score.value() < threshold.value() as f64;
 
     println!("Architecture Compliance CI");
     println!("Score: {:.1} / 100", score.value());
